fix(test): regenerate scene165 golden per-run (force) for cross-platform parity
scene165 parity failed on the cloud browser (MAD 0.275) because the golden was
captured on Windows and compared against the cloud-rendered Lite output. Capture
the BJS oracle golden with force:true so CI regenerates it on the same machine,
and merge the culling test into the same serial spec so the shared golden is
never written concurrently.
